Pushing some changes for basic doxygen file setup.
adding and changing the testing files.

Updated the Catch2 to be on version 2.13.4 until version 3 is stable. Got a semi hacky test in for now.

The instruction for CMake integration on Catch2 were failing. I am not sure why but spent a good chunk of time but couldn't figure it out.

Testing is done by using the new test command to compile and run the tests.