Split and moved files
I separated the header and source. If i want to have a header only version i think scripting to get there is best.

It does compile a bit faster now. I have the source files named for now in the cmake. I might make that auto discover but not sure yet.